myapp/serializers.py
from rest_framework import serializers
from . import models


class UserSerializer(serializers.ModelSerializer):

    class Meta:
        fields = ('first_name', 'last_name', 'email', 'birthday',)
        model = models.User
